@@ -7,20 +7,20 @@ readme = "README.md"
 packages = [{include = "src"}]
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.13"
 pyotp = "^2.9.0"
 opencv-python = "^4.8.1"
 cryptography = "^41.0.7"
-Pillow = "^10.0.1"
-docker = "^6.1.3"
 numpy = "^1.24.0"
 
-[tool.poetry.group.dev.dependencies]
+[tool.poetry.group.test.dependencies]
 pytest = "^7.0.0"
 pytest-cov = "^4.0.0"
 pytest-mock = "^3.10.0"
 pytest-asyncio = "^0.21.0"
-black = "^23.0.0"
+
+[tool.poetry.group.lint.dependencies]
+black = "^24.0.0"
 flake8 = "^6.0.0"
 mypy = "^1.0.0"
 
@@ -33,10 +33,10 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 88
-target-version = ['py38']
+target-version = ['py312']
 
 [tool.mypy]
-python_version = "3.8"
+python_version = "3.13"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
