@@ -4,4 +4,6 @@
 
 __version__ = "1.0.0"
 __author__ = "OneTimePassword Team"
-__description__ = "Google Authenticatorと同様の機能を持つワンタイムパスワード生成アプリケーション"
+__description__ = (
+    "Google Authenticatorと同様の機能を持つワンタイムパスワード生成アプリケーション"
+)
